Feat: 플레이 수정

+++ code/FourQuest/FQClient/DeadArmour.h
@@ -28,6 +28,6 @@ namespace fq::client
 	private:
 		game_module::PrefabResource mLivingArmour;
 
-		unsigned int mPlayerCount;
+		unsigned short mPlayerCount;
 	};
 }
